[#735] remove dependency on test folder
Moved function definitions for make_session and home_collection to irods.helpers module.

Made irods.test.helpers versions of those functions import and/or derive from those definitions.
