+++ irods/test/helpers.py
@@ -1,25 +1,29 @@
-import os
-import io
-import tempfile
-import contextlib
-import shutil
-import hashlib
 import base64
+import contextlib
+import io
+import datetime
+import hashlib
+import inspect
+import json
 import logging
 import math
+import os
+import shutil
 import socket
-import inspect
-import threading
 import random
-import datetime
-import json
+import re
 import sys
+import tempfile
+import threading
+
 import irods.client_configuration as config
-import irods.rule
-from irods.session import iRODSSession
+from irods.helpers import (
+    home_collection,
+    make_session as _irods_helpers_make_session)
 from irods.message import iRODSMessage, IRODS_VERSION
 from irods.password_obfuscation import encode
-from irods import env_filename_from_keyword_args
+import irods.rule
+from irods.session import iRODSSession
 
 
 class iRODSUserLogins:
@@ -74,19 +78,6 @@ def configuration_file_exists():
     except config.NoConfigError:
         return False
     return True
-
-
-class StopTestsException(Exception):
-
-    def __init__(self, *args, **kwargs):
-        super(StopTestsException, self).__init__(*args, **kwargs)
-        if "unittest" in sys.modules.keys():
-            print("Aborting tests [ Got : %r ]" % self, file=sys.stderr)
-            os.abort()
-
-
-class iRODS_Server_Too_Recent(StopTestsException):
-    pass
 
 
 def my_function_name():
@@ -172,48 +163,13 @@ def make_environment_and_auth_files(dir_, **params):
     return (config, auth)
 
 
-def get_server_version_for_test(session, curtail_length):
-    return session._server_version(session.GET_SERVER_VERSION_WITHOUT_AUTH)[
-        :curtail_length
-    ]
-
-
-# Create a connection for test, based on ~/.irods environment by default.
-
-
 def make_session(test_server_version=True, **kwargs):
-    """Connect to an iRODS server as determined by any client environment
-    file present at a standard location, and by any keyword arguments given.
-
-    Arguments:
-
-    test_server_version: Of type bool; in the `irods.test.helpers` version of this
-                         function, defaults to True.  A True value causes
-                         *iRODS_Server_Too_Recent* to be raised if the server
-                         connected to is more recent than the current Python iRODS
-                         client's advertised level of compatibility.
-
-    **kwargs:            Keyword arguments.  Fed directly to the iRODSSession
-                         constructor."""
-
-    env_file = env_filename_from_keyword_args(kwargs)
-    session = iRODSSession(irods_env_file=env_file, **kwargs)
-    if test_server_version:
-        connected_version = get_server_version_for_test(session, curtail_length=3)
-        advertised_version = IRODS_VERSION[:3]
-        if connected_version > advertised_version:
-            msg = (
-                "Connected server is {connected_version}, "
-                "but this python-irodsclient advertises compatibility up to {advertised_version}."
-            ).format(**locals())
-            raise iRODS_Server_Too_Recent(msg)
-
-    return session
+    return _irods_helpers_make_session(test_server_version=test_server_version, **kwargs)
 
 
-def home_collection(session):
-    """Return a string value for the given session's home collection."""
-    return "/{0.zone}/home/{0.username}".format(session)
+make_session.__doc__ = re.sub(
+    r"(test_server_version\s*)=\s*\w+", r"\1 = True", _irods_helpers_make_session.__doc__
+)
 
 
 def make_object(session, path, content=None, **options):
